Update index.html
Column set for the nav, main, and aside. The bootstrap grid is always a 12 grid layout. Therefore, the nav was set to 3 to give it a 1/4 of the page length, the main was set to 6 to give it 1/2 of the page length, and the aside was give 3 to also give it 1/4 of the page length.